requirements.txt
python-dotenv==0.19.0
boto3==1.18.5